+++ src/lib/services/txService.ts
@@ -16,7 +16,7 @@ import {
   getBlockTransactionsByHeightQueryDocument,
 } from "lib/query";
 import { createQueryFnWithTimeout } from "lib/query-utils";
-import type { Addr, Option, Transaction, TxFilters, Message } from "lib/types";
+import type { Addr, Option, Transaction, TxFilters } from "lib/types";
 import { MsgFurtherAction } from "lib/types";
 import {
   getActionMsgType,
@@ -89,7 +89,7 @@ export const useTxsByAddressPagination = (
       .then<Transaction[]>(({ account_transactions }) =>
         account_transactions.map<Transaction>((transaction) => ({
           hash: parseTxHash(transaction.transaction.hash),
-          messages: snakeToCamel(transaction.transaction.messages) as Message[],
+          messages: snakeToCamel(transaction.transaction.messages),
           sender: transaction.transaction.account.address as Addr,
           isSigner: transaction.is_signer,
           height: transaction.block.height,
@@ -207,7 +207,7 @@ export const useTxs = (
         .then<Transaction[]>(({ transactions }) =>
           transactions.map<Transaction>((transaction) => ({
             hash: parseTxHash(transaction.hash),
-            messages: snakeToCamel(transaction.messages) as Message[],
+            messages: snakeToCamel(transaction.messages),
             sender: transaction.account.address as Addr,
             isSigner: false,
             height: transaction.block.height,
@@ -267,7 +267,7 @@ export const useTxsByBlockHeightPagination = (
         .then<Transaction[]>(({ transactions }) =>
           transactions.map<Transaction>((transaction) => ({
             hash: parseTxHash(transaction.hash),
-            messages: snakeToCamel(transaction.messages) as Message[],
+            messages: snakeToCamel(transaction.messages),
             sender: transaction.account.address as Addr,
             isSigner: false,
             height,